+++ modbus_configuretools/__init__.py
@@ -1,3 +1,4 @@
 from . import modbus_wrapper 
 from . import derived_modbus_wrapper 
 from . import temperature_sensor 
+from . import helper_function
